feat(frontend): @agenta/observability — the analytics query, range and transform leave the app
