+++ web/packages/agenta-observability/package.json
@@ -1,0 +1,42 @@
+{
+    "name": "@agenta/observability",
+    "version": "0.1.0",
+    "private": true,
+    "sideEffects": false,
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
+    "scripts": {
+        "build": "tsc --noEmit",
+        "types:check": "tsc --noEmit",
+        "lint": "eslint --config ./eslint.config.mjs src/",
+        "test": "pnpm run test:unit",
+        "test:unit": "vitest run",
+        "test:watch": "vitest",
+        "check": "pnpm run types:check && pnpm run lint"
+    },
+    "exports": {
+        ".": "./src/index.ts",
+        "./state": "./src/state/index.ts"
+    },
+    "dependencies": {
+        "@agenta/entities": "workspace:../agenta-entities",
+        "@agenta/shared": "workspace:../agenta-shared",
+        "dayjs": "^1.11.20"
+    },
+    "peerDependencies": {
+        "@tanstack/react-query": ">=5.0.0",
+        "jotai": ">=2.0.0",
+        "react": ">=18.0.0",
+        "jotai-tanstack-query": ">=0.9.0"
+    },
+    "devDependencies": {
+        "@tanstack/react-query": "5.100.9",
+        "@types/node": "^20.19.20",
+        "@types/react": "^19.0.10",
+        "jotai": "^2.15.0",
+        "react": "^19.0.0",
+        "typescript": "^5.9.3",
+        "vitest": "^4.1.4",
+        "jotai-tanstack-query": "^0.11.0"
+    }
+}
